query first data from DB
DB and resolver are finally hooked together. Initially I also added ratings service to fetch ratings based on book ID, but that isn't needed when `with` drizzle option is used, so I removed it.

To make magic `with` option work, I had to add all relations to the schema. I also had to add a hack to convert enum to pgEnum, because drizzle-orm's typescript definitions are not perfect yet.

This commit also marks an important decision moment. From now on, schema source of truth is in `schema.ts` file (see how book model imports `Genre` from schema). Reason is, I want to keep schema definition as close to DB as possible. DB is, after all, physical representation of the schema and I don't want definitions from services or models to flow all the way down to the DB.