add authentication
I opted for `passport` and `jwt`. Passport because it is very versatile for the future - we can switch to OAuth in no time. But for now, `jwt` is sufficient. Well, I guess current implementation is super frustrating for consumer, because auth token is invalidated every 60 seconds and you literally have to `login` with username and password again. I am aware that proper implementation needs long-living tokens, refresh tokens, etc. But current implementation is more than enough for the demo.

The same applies for token persistency. It lives in the memory which is a nice quick hack, but any proper product needs to store these in the DB - to ensure every product update doesn't log out all users. And to allow scaling the app horizontally.

Please note that I didn't implement registration or user management in any way. For now, we have two hardcoded users. User management will come in next commit.