add pagination
There are many types of paginations actually. Endless scrolling, page numbers, etc. For the simplicity of solution I opted for "endless scrolling" pagination that is actually very, very easy. All you need is ability to set `offset` to query that returns arrays. On frontend side you just call the same query with bigger offset and if response is an empty array, you got to the end of the list.

Another argument for "endless scrolling" pagination? You don't force DB to compute total number of items which may become too expensive for large datasets. Yes, we can store that number in cache, but purpose of this project isn't that.

And to be honest, I go myself into a trap by making `reviews` table. As you can see, you can paginate between books and history of the book, but you cannot paginate between reviews of the book. That's because drizzle doesn't support offset of sub-level queries. I know it is stupid, but I left it the way it is because adding offset to sub-query would mean rewriting whole query from scratch and I evaluated it is not worth the time.