update docs
To help newcomers start with the project, it is always a good idea to have proper README set in place. I kept the original version, and added "real" README below it. It is always difficult to write "Getting Started" section when you already have everything installed, I can just hope it is alright and receive feedback from people who are actually trying to run the project.